requirements.txt
aiofiles==24.1.0
aiohttp==3.9.0
boto3==1.35.12
crossplane==0.5.8  #On Boarding requirements
ddapm-test-agent==1.18.0
dictdiffer==0.9.0  # for parametric tests
docker==7.1.0
filelock==3.12.2  # for parametric tests
jsonschema==4.16.0
kubernetes==29.0.0  #lib-injection kubernetes
matplotlib  # for performance tests
mitmproxy==9.0.1
mock==5.2.0
msgpack==1.0.4
mypy==1.15.0
opentelemetry-api==1.27.0 # for parametric tests
paramiko==3.4.1
pexpect==4.9.0  #On Boarding requirements
pluggy==1.0.0
psutil==6.0.0
pulumi-aws==6.38.0  #On Boarding requirements
pulumi-command==0.11.1  #On Boarding requirements
pulumi-tls==5.0.3  #On Boarding requirements
pulumi==3.116.0  #On Boarding requirements
pytest-json-report==1.5.0
pytest-split==0.9.0
pytest-xdist==3.2.1
pytest==7.1.3
python-dateutil==2.8.2
python-vagrant==1.0.0  #On Boarding requirements
pyyaml-include==1.3  #On Boarding requirements
pyyaml==6.0.1  #On Boarding requirements
requests==2.31.0
retry==0.9.2
rfc3339-validator==0.1.4
ruff==0.14.5
scp==0.14.5
semantic-version==2.10.0
setuptools==75.8.0
shellcheck-py==0.11.0.1
types-aiofiles==24.1.0.20241221
types-mock==5.2.0.20250924
types-protobuf==5.29.1.20241207
types-python-dateutil==2.9.0.20241206
types-PyYAML==6.0.12.20241230
types-requests==2.31.0
types-retry==0.9.9.20250322
watchdog==3.0.0
yamllint==1.35.1
yarl==1.9.4
